+++ src/Groups/GroupsList/index.js
@@ -1,2 +1,1 @@
-export { default as GroupsList } from './GroupsList';
 export { default as GroupsListContainer } from './GroupsListContainer';