+++ test/bigtest/network/factories/fundType.js
@@ -2,6 +2,6 @@ import { Factory } from 'miragejs';
 import faker from 'faker';
 
 export default Factory.extend({
-  id: faker.random.uuid,
+  id: faker.datatype.uuid,
   name: faker.finance.accountName,
 });