+++ test/bigtest/network/factories/ledger.js
@@ -4,7 +4,7 @@ import faker from 'faker';
 import { LEDGER_STATUS } from '../../../../src/Ledger/constants';
 
 export default Factory.extend({
-  id: faker.random.uuid,
+  id: faker.datatype.uuid,
   name: faker.finance.accountName,
   code: 'ONETIME',
   description: faker.random.word,